move exception classes to a sub-package, otherwise this will get messy
